pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "srag"
version = "0.4.0"
description = "Smart RAG — local Tavily-style search + scrape + vector index for LLM/RAG workflows"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "trafilatura>=1.8",
    "duckduckgo-search>=6.0",
    "lancedb>=0.6",
    "sentence-transformers>=3.0",
    "pyarrow>=15.0",
    "python-dotenv>=1.0",
]

[project.scripts]
srag = "srag.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["srag*"]